+++ nessai/flows/__init__.py
@@ -6,14 +6,14 @@ implementations.
 from .base import BaseFlow, NFlow
 from .maf import MaskedAutoregressiveFlow
 from .nsf import NeuralSplineFlow
-from .realnvp import FlexibleRealNVP
-from .utils import setup_model, reset_weights, reset_permutations
+from .realnvp import RealNVP
+from .utils import configure_model, reset_weights, reset_permutations
 
 __all__ = ["BaseFlow",
            "NFlow",
            "MaskedAutoregressiveFlow",
            "NeuralSplineFlow",
-           "FlexibleRealNVP",
-           "setup_model",
+           "RealNVP",
+           "configure_model",
            "reset_weights",
            "reset_permutations"]